[babel] Add missing babel-plugin dependencies
`babelRegisterOnly.js` requires several Babel plugins that are listed as devDependencies in this repo's root package.json but aren't listed as dependencies in the metro-bundler package's dependencies. If you install metro-bundler in another project and require `babelRegisterOnly`, the Babel plugins are missing at runtime. This adds those missing Babel plugins to package.json.