Add an option to enable/disable .babelrc lookup
By default, when Babel transforms a file, it looks for .babelrc in the file's directory and all parent directories until it finds a .babelrc file. This means that when Metro tries to transform `<PROJECT_ROOT>/node_modules/dep/mod.js`, it searches for `<PROJECT_ROOT>/node_modules/dep/.babelrc`. If that .babelrc actually exists, Babel will try to apply it.

In practice, this often causes problems because packages that include .babelrc often do so unintentionally -- they don't intend for the package consumer to look at that .babelrc file. One thing we've done a lot is `rm` all .babelrc files under node_modules -- this commit effectively achieves the same in a less destructive way by telling Babel not to look up .babelrc files. (To clarify, Metro will still apply the .babelrc file in the project root.)

Since the current behavior is to look at .babelrc files, this commit keeps that behavior for now. We'll consider overriding the default behavior (that is, making Babel not lookup .babelrc files) in the default configuration of Expo projects. If this goes well and we empirically find that people are having a better time, we may want to consider flipping this option's default in Metro, so that Metro tells Babel not to look up .babelrc files by default.

Test Plan: Made the corresponding changes to react-native to add this option to rn-cli.config.js. Then made a dummy .babelrc file in react-native's node_modules, where the bundler would eventually look:

```
$ cat node_modules/react/.babelrc
{
  "plugins": ["dummy"]
}
```

Loaded the RNTester project with `enableBabelRCLookup` set to true and got a redbox saying `Unknown plugin "dummy"`. Changed the config on the RN side to set `enableBabelRCLookup` to false, cleared the packager cache, and reloaded the bundle successfully.