Expose createModuleIdFactory as bundler option
This enables `metro.runBuild()` to be run with a custom module ID factory.

One use case is building a base bundle, on top of which other application-specific bundles could reference modules in the base.  The application-specific IDs need to not conflict with those in the bae bundle, and all references to modules in the base must resolve to the correct ID in the base bundle.  A custom ID factory can make all this possible.