metro: Package: fix/simplify replacement behavior for string `browser` field
Summary:
The spec doesn't say the `browser` field should actually generate a replacement for the main *file*, just that it provides an alternative main: https://github.com/defunctzombie/package-browser-field-spec#alternate-main---basic
We also don't have any test covering this behavior. Any package relying on this behavior should be fixed to have a redirection of the main *file* in addition to their `main` *field*.

This is part of an effort to cleanup the way we do redirections. Eventually this will be part of the resolution module, not `Package.js`.

Reviewed By: davidaurelio

Differential Revision: D6611894

fbshipit-source-id: 0186aa23df2a9183d895ea02b5fd3b2b7339dc76