Tweak circleci configuration
We've had the node v6 test step failing for some time now (due to https://github.com/prettier/prettier/issues/3457).

Since there seems to be no progress on that issue, I'm adopting a similar approach that React Native has done to work around the issue: run the linter and flow on the latest nodejs version, and only run the jest tests on previous versions (this is also that jest seems to be doing as well).

For this I've created a run-js-checks step, while the jest tests are still under the test-node-X steps