Fix Metro sourcemaps when minification is enabled
Summary:
In the old days, the module wrapping was done by just prepending a string to the module source code. I guess that in order to fix the sourcemaps, an offset was added. This is not needed anymore (and in fact is causing issues when generating the sourcemaps).

This should fix https://github.com/facebook/react-native/issues/6946

Reviewed By: davidaurelio

Differential Revision: D8223377

fbshipit-source-id: 629a6bf69875a3faa25c79fb2744727dedfe0924