Updates to UnableToResolveError to account for different platforms.
This error message is confusing when you get it on a mac, as /tmp/ isn't the
tmp directory that is being used.  The message should probably try to account
for settings that are configured by users as well, the Metro Bundler cache at
least is something that can be overwritten.