Fix entryFile always is assumed to end with `.js` extension (#310)
Summary:
<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

**Summary**
When making a request to metro server with `.bundle` suffix, if this source file was named with `.jsx` or `.ts`, metro isn't able to find the resource. Instead, an error message is shown indicating there isn't such file associated with `.js` extension.

This PR addresses this issue by passing the entryFile name in the URL into the module resolution system, which will handle various extensions gracefully.

**Test plan**
- Test in the application, making a request to a source file with `.jsx` extension.
Pull Request resolved: https://github.com/facebook/metro/pull/310

Reviewed By: rafeca

Differential Revision: D12981415

Pulled By: mjesun

fbshipit-source-id: d0b592800f6bdec1850149165d9e925ad270485a