Update to gh-pages Fixes Tagline and Home Page Image & Fix for Favicon
**Summary**

Added the  "🚇"  to the title and Home Page name of facebook.github.io/metro/". Before I added it, it was showing to the public a Vertical Rectangle sign. Now it's showing the correct favicon  "🚇

When @mjesun came to Lagos,Nigeria  he talked so passionately about how Facebook is building Metro to help build faster Apps, so I decided to take a look at getting started with using Metro-Bundler.

**Test plan**

Once its merged the Tagline Favicon would should on the website at facebook.github.io/metro/ as "🚇 The JavaScript bundler for React Native" and also on the Home Page Header.

## How It looks on my Fork with my Changes

<img width="124" alt="metro2" src="https://user-images.githubusercontent.com/20538832/49222809-c29e3580-f441-11e8-9e21-5ea966c04a68.PNG">

<img width="614" alt="metro" src="https://user-images.githubusercontent.com/20538832/49222810-c29e3580-f441-11e8-9ee9-dc76af7f36b6.PNG">


## How it looks currently of your End

<img width="130" alt="metro_old1" src="https://user-images.githubusercontent.com/20538832/49222877-03964a00-f442-11e8-984e-f7f0af207063.PNG">

<img width="600" alt="metro_old" src="https://user-images.githubusercontent.com/20538832/49222878-03964a00-f442-11e8-864b-951f562cdd1c.PNG">