Record class names in function map
Summary: In the case of a class with an default constructor, stack traces may point to the first line of a class declaration (e.g. if that class extends `Error`). There are other possible cases of this, for example in property initialisers. This diff adds class names to the function map so we can usefully symbolicate such stack frames, instead of falling back to the name of the surrounding function (usually `<global>`.

Reviewed By: cpojer

Differential Revision: D16650752

fbshipit-source-id: 6eb6d25e81ea9757d95718eaab09822d9bd25ab8