Pass the original module name to custom resolver
We are using a custom resolver for our custom module resolution logic,
and we got an issue where the module name passed to our custom resolver
is not the actual name of the module in the node-modules.
Our investigation find that the metro's resolver will change the original
module name as per the redirect defined in "react-native" or "browser"
field of the package.json of that requiring package.

I found that, by just adding the original module name to the list of
arguments passed to the custom resolver will solve the problem without
affecting any of its existing users. Also this will give the custom
resolver a better context on the required module, which the custom
resolver can handle in a better way.