Handle circular dependency upon dependency removal (#506)
Summary:
**Summary**
This is a proposed fix for the issue I submitted https://github.com/facebook/metro/issues/500. For problem statement, please refer to the issue linked to understand the scope of the bug.

Currently Metro does not handle dependency removal correctly in the presence of certain circular dependency. Since we rely on Metro server to produce correct dependency graph in hopes to bundle only the needed modules. This is especially an issue when a dependency is removed from the entry point, however, due to it having circular dependency, the updated dependency graph will leave out all its dependencies in the graph.

I added the following two more test cases to ensure the correctness of the algorithm:

1. Remove `B` from `E`: `removes a dependency with transient cyclic dependency`
![image](https://user-images.githubusercontent.com/1754430/71753699-39bdae80-2e38-11ea-82f7-ab4285acc869.png)

2. Remove `B` from `E`: `removes a cyclic dependency which is both inverse dependency and direct dependency`
![image](https://user-images.githubusercontent.com/1754430/71753852-bea8c800-2e38-11ea-9166-4f97854b1faa.png)

3. Remove `B` from `E`: `removes a sub graph that has internal cyclic dependency`
![image](https://user-images.githubusercontent.com/1754430/71753928-ffa0dc80-2e38-11ea-8c9e-cdd292290472.png)

Feel free to propose more tests to ensure the correctness of the algorithm. I'm aware that this may introduce more expensive graph updates for certain scenarios, but I believe that ensuring the correctness is far more important for dependency graph updates.

## Implementation
Previously, our implementation will stop removing circular dependency due to it having remaining `inverseDependencies`:

https://github.com/facebook/metro/blob/984aab84358aeef8de13ba86c33834a1c1bf6926/packages/metro/src/DeltaBundler/traverseDependencies.js#L328-L330

This can be illustrated by this example:
![image](https://user-images.githubusercontent.com/1754430/71753699-39bdae80-2e38-11ea-82f7-ab4285acc869.png)

When removing `B` from `E`, `B` will still have an inverse dependency `A` left. Henceforth, the rest of the graph remain untouched. My proposed solution is to have this async function `canSafelyRemoveFromParentModule` recursively checking all the inverse dependencies. In this example, we will look up inverse dependencies of `A` all the way to the end until there is no inverse dependency. We can only safely remove this dependency if and only if its end inverse dependency (in this case `A` will have `B` as its end inverse dependency) only has one path and the path is the same as the parent path.

**Test plan**
- Updated 1 existing test, with 2 additional unit tests to ensure the correctness of the dependency removal logic handling with circular dependency.

```
yarn run jest packages/metro/src/DeltaBundler/__tests__/traverseDependencies-test.js
```
Pull Request resolved: https://github.com/facebook/metro/pull/506

Differential Revision: D19533899

Pulled By: cpojer

fbshipit-source-id: 80ea0d0413574cc7e7f85f84529ca948d938861c