add ios/android build dirs to blacklist
On windows rebuilding an app while metro is running causes both the build and metro to crash. 
This means that every time you rebuild your native code you need to restart metro and wait for it to complete bundling, which can be a nightmare when debugging native code.
This fixes this issues.