Create metro-transform-plugins package
Summary:
For some reason all these plugins are in the `JSTransformer/worker` folder. This diff moves all of them out of the main Metro package and into a separate package. I will likely add other things to this package that fit there later.

This is in preparation to use HBC in Metro. The next step is to rename `JSTransformer/worker` and clean it up.

Reviewed By: motiz88

Differential Revision: D20955820

fbshipit-source-id: 9596e189e1500792799e0ab2ce4365d72c5b5c28