Minor cleanup of Babel types
Summary:
The Babel types are a mess. This diff cleans up the definitions (a little bit, there is tons more work to be done to make them actually good), removes some FlowFixMe's, and fixes the definition for the template function.

Note: I am doing this to unblock a diff and I couldn't keep working until I cleaned up the file a little bit. I know it is not perfect but I'm not looking to spend a lot of time on it right now.

Reviewed By: rubennorte

Differential Revision: D21089593

fbshipit-source-id: e1498ad7a45579f76e3ef40026eff3ff6d48e234