One more upgrade/cleanup pass of third-party deps
Summary: Every time after dropping a large amount of deps or aligning their versions the layout and remaining dependency tree changes enough that it makes sense to go through all the deps again and see which ones can be removed. This is just yet another manual pass of that to bring a few things to the latest versions and remove some that are now unreferenced.

Reviewed By: yungsters

Differential Revision: D21428538

fbshipit-source-id: 791155a0e7412a68851af5e4df056080dbd456e1