Align state property naming
Summary:
The goal of this stack is to unify the Metro core implementation of `collectDependencies` with an internal fork.

This diff renames the `dependency` state property of the metro implementation to `nextDependencyIndex` to align it with the internal fork.

This isn't strictly necessary but makes it easier to reason if the two implementations do the same and I find `nextDependencyIndex` the more meaningful name.

Reviewed By: motiz88

Differential Revision: D23472496

fbshipit-source-id: 9f9a5f0238f5af67ca4a7ac4c09f3a15176db663