Merge the dependencyData and indices state
Summary:
The goal of this diff and follow up diffs is to align the metro `collectDependencies` implementation with an internal fork.

This diff merges the `dependencyData` and the `dependencyIndexes` maps to bring the implementations closer together. It has the nice . side effect that it reduces the number of required map lookups and use of `nullthrows`.

Reviewed By: motiz88

Differential Revision: D23539114

fbshipit-source-id: 3d599d5934d5cbba2acaa9ff4e84b3496dbba480