Support custom module dependency registry
Summary:
The goal of this stack is to unify the `collectDepenencies` implementation with an internal fork.

The `collectDependencies` implementation of the internal fork and metro-core treat dependencies differently:

* metro-core: Combines all dependencies with the same relative path into one dependency.
* internal-fork: Returns each dependency separately.

Ideally, the `collectDependencies` function should return a separate dependency for every `asyncType` and optionally to e.g. enable precise error reporting. However, the `traverseDependencies` implementation identifies modules mainly on the path and does not respect different `asyncTypes` (a `prefetch` and `async` dependency are accidentally merged).

This diff adds support for a custom module dependency registry that decides when dependencies should be collapsed per module. A following diff will add support to parametrize the registry to support the different behavior of `metro-core` and the internal fork.

Reviewed By: motiz88

Differential Revision: D23539266

fbshipit-source-id: 6388c664ed86bf16e51b71d900006017029d269a