Refine types for @babel/traverse
Summary:
This diff refines/completes the types for babel/traverse.

I started with the type definitions from Typescript and then added missing methods and refined the typing because the Typescript definitions are somewhat outdated (more than a year old).
I tried to infer the correct types but there's a chance that a type is more restrictive/generic than it should be. I believe this isn't a problem as we can change the definition if they show to be incorrect.

The `NodePath` definition isn't sound because the declaration defines most fields as non-nullable but the fields are all initialized with `null` after calling the constructor only. Typing all the fields as `nullable` makes it fairly cumbersome to use the type in `Visitors` which is the main use case and babel/traverse ensures that `NodePath.setup` is called for every node which ensures that the fields are initialized.

We can contribute these types to flowtyped if my PR for `babel/types` is accepted.

Reviewed By: cpojer

Differential Revision: D24220279

fbshipit-source-id: d2f504ad38b64c15db1c2a803b88294f932de298