Fix resolver ignoring `/node_modules`
Summary: I noticed that our resolver code has a bug in that it looks for `node_modules` at `/Users/node_modules` but not in `/node_modules`. This is a bug, as the spec outlines that every directory up to the root has to be considered (See `NODE_MODULES_PATHS` in https://nodejs.org/api/modules.html#modules_all_together).

Reviewed By: motiz88

Differential Revision: D24300607

fbshipit-source-id: a56aa0bc4e83e998fa5fff3bc9dcfb605e5f27e8