Remove `projectRoot` from cache keys
Summary:
**tl;dr: Remote cache is broken, working on fixes, this one removes a legacy cache key value.**

The remote cache has been broken since I rolled out the Metronome/DotSlash package of Metro. I believe the reason is now due to a cache key mismatch but I don't have good insight into the actual cache key calculation on Sandcastle since the key we are using to write into the cache is not the same one we display on the job. The reason for that is because we use a prebuilt version of Metro but we run the cache-filling script from source. While I'm working on merging this script into `fb-tools/fb-metro-cli` to get more insights into cache keys, I'm trying to fix the issue based on my investigation into the source code of Metro.

For this particular diff, I noticed that we add the relative path of the project root into the cache key. This does not make any sense as we are calculating a relative path from the project root to Metro, therefore tightly coupling the location of Metro on the system and the configuration Metro uses to run code. For example, when run from source, Metro produces `../../..` as a value and `../../../../../../../../../fbsource/xplat/js` as a value when run via Metronome on my MacBook. I suspect that this path may be different on Sandcastle and therefore leading to cache misses.

If all configuration in Metro, including the transformer, are the same for two Metro setups - which is highly unlikely - but even if, and the file that is transformed has the same sha1 across projects, then we can also safely re-use the cache as the concrete location in such a case shouldn't matter.

I traced the introduction of this relative path back to D4237840 (https://github.com/facebook/metro/commit/2322db4f819fde85bcb8867e1ef6ceb860b15ef4) and arguably even then we should have considered doing something differently.

Reviewed By: sammy-SC

Differential Revision: D24853373

fbshipit-source-id: dc25d7378d72c07649cced78fe4d59dc3eb9e5d0