Fix worker test (#619)
Summary:
**Summary**

7667838d716a032b77f16187887170358672b95d presumably fixed the worker-test but I forgot to copy-paste the `console` mock fix when creating the diff in the FB internal repo. Applying the changes now and creating a public PR to ensure the CircleCI passes before merging.

I further noticed that tests can occasionally fail because of race conditions. The `end` method listens on the `end` event which is fired before the `finish` event on the stream. That means that the streams of a test might not yet have been closed when the `afterAll` asserts that all are closed.

This diff changes the `openStreams` to promises which can be awaited in the `afterAll`. The alternative would have been to wait on the `finish` event in the `end` function but that would have required changes to production code which I rather not make.

**Test plan**

`yarn test`

Pull Request resolved: https://github.com/facebook/metro/pull/619

Reviewed By: cpojer

Differential Revision: D25517485

Pulled By: MichaReiser

fbshipit-source-id: e20d21bab761b6e6e3b02a187f751cc67daec9f1