Extract Asset transformation in transform-worker
Summary:
This diff is part of a stack that re-structures the metro-transform-worker with the goal to make the transformation pipeline(s) more visible as well as what metadata is added by each step. The ultimate goal is to model the pipeline as a set of transformation-steps where the output of one step is the input of the next step (e.g. the asset step transforms an asset to a JS artifact that than is transformed by the JS step).

It extracts the `asset` transformation into its own function that first transforms the asset to JavaScript and then calls into the JS transformation.

Reviewed By: motiz88

Differential Revision: D28386005

fbshipit-source-id: 58ca2b27aba09636e6a838c452244b40f8033ac3