Resolve package.json module replacements when the replaced module isn't in ModuleCache
Summary:
Adds a `getPackageOf(modulePath)` method to the Metro resolver's internal `ModuleCache` interface. This method can be used instead of `getModule(modulePath).getPackage()` in cases where `modulePath` itself may not be in the cache (or the filesystem) but its `package.json` is.

This specifically applies when [replacing modules in `browser`-like-fields in `package.json`](https://github.com/defunctzombie/package-browser-field-spec#replace-specific-files---advanced): if the "intermediate" path (that is being redirected to something else) doesn't correspond to a module in the cache, resolution currently fails, but it should succeed because the intermediate module's contents/existence cannot affect the result of the redirection. Here, we use `getPackageOf()` to represent this intent and fix this case.

Reviewed By: mdvacca

Differential Revision: D30776655

fbshipit-source-id: 9355ec08214124634be43996d73889be58c11ee9