fix(metro-config): Pass absolute default worker path to metro
If we pass an unresolved path to metro, metro's worker pool,
jest-worker, will need to resolve the worker path itself. This relies on
jest-worker being able to resolve metro, even though it doesn't
explicitly depend on metro (doing so would result in a nice cyclic
dependency).

In package managers that eliminate phantom dependencies, like Yarn PnP
or pnpm, this resolution just fails.

To avoid relying on a phantom dependency, we need to resolve the
absolute path of the worker before handing it off to jest-worker. Since
_we_ depend on metro, this is fine, and jest-worker can consume the
absolute path.