Remove module resolution troubleshooting steps
Summary: Removes the "If you are sure the module exists, try these steps" part of the error message shown when Metro fails to resolve a module. In almost all cases, this message is a red herring and causes needless confusion. When it appears, the real issue is more likely along the lines of a misspelled file name somewhere.

Reviewed By: GijsWeterings

Differential Revision: D32095928

fbshipit-source-id: 7e21b45ba00caf76312b5efd43382017812e57f1