Added named groups babel transform to rn babel preset
Hermes currently does not support regex named groups. Whenever a library is included that uses named groups the rn build with hermes will fail (as it doesn't compile). This is already the case when using jest 27 as this includeds a dependency (ansi-styles) that uses named-groups which is transitivily included via pretty-format in HMRClient