Add explicit dependency on `@babel/runtime` to `metro-runtime`
Summary:
Previously, `babel/runtime` was a transitive dependency of RN projects via `metro-react-native-babel-preset` -> `babel/plugin-transform-regenerator`.

`babel/runtime` is a dependency of injected babel helpers, so we need it to be a transitive dependency of an RN project under `react-native`.

Removing the previous implicit dependency broke RN builds, this adds the dependency to `metro-runtime` to make it explicit.

I've created a 0.70.x branch for Metro from 0.70.2 and will cherry pick this up as a hotfix 0.70.3 release.

Reviewed By: motiz88

Differential Revision: D36165366

fbshipit-source-id: d7e13cf337acb88e9d9733f8caf332aa4303ab97