Mark `metro-file-map` as experimental, rename `hasteMapCacheDirectory` -> `fileMapCacheDirectory`
Summary:
`metro-file-map` has not yet been part of an OSS Metro release. This marks it as "experimental" (i.e. semver-excluded), for as long as we're making regular changes to its API, which is typically only consumed by other Metro packages.

Additionally, this makes a top-level config name change `hasteMapCacheDirectory` -> `fileMapCacheDirectory`, to reflect `jest-haste-map` -> `metro-file-map`, while retaining support for `hasteMapCacheDirectory` as a deprecated alias.

Reviewed By: motiz88

Differential Revision: D36451806

fbshipit-source-id: 498271c1b4b6037cb567b5f87abee2f47b094916