Add missing .npmignore files
Summary:
I noticed some of our packages publish `src.real` to NPM, which is an intermediate build artefact that shouldn't published. It turns out this is because a few of them are missing the standard `.npmignore`, and so also publishing test code and lock files.

This adds the missing `.npmignore` files, and adds a check to our `subpackages-test.js` for existence and minimal content.

Reviewed By: motiz88

Differential Revision: D36592818

fbshipit-source-id: adac01c5c9d632e3388290a46c3cbc3308b4e239