feat: add support for capturing require.context in dependency collector (#821)
Summary:
> First part of my [`require.context`](https://webpack.js.org/api/module-methods/#requirecontext) feature request. [Learn more](https://github.com/microsoft/rnx-kit/discussions/1515).

The key here is to add support for importing the contents of a folder recursively and providing to the user as some sort of map, the regular expression is an added layer for refining the results. This means in areas where we'd register a filename, we'd use a directory name instead.

This PR implements the ability to capture `require.context` invocations in `collectDependencies`. One could think of this like "language support" for the `require.context` feature, in the sense that Metro can now make sense of the new syntax.

Because multiple dependencies can use the same module identifier (`require.context('[module-id]')`) but also provide extra properties like `recursive`, or a RegExp matcher, we need to change the registry key format to include extra properties. The aforementioned mentioned functionality is primarily implemented in `getKeyForDependency`.

The entire feature is implemented behind a feature flag `unstable_allowRequireContext`, e.g.

`metro.config.js`
```js
const metroConfig = {
  // ...
  transformer: { unstable_allowRequireContext: true }
};
```

This PR does not implement the entire `require.context` feature, we'll need a follow up PR implementing changes to the `DependencyGraph`, `DeltaBundler`, and `metro-runtime`.

Pull Request resolved: https://github.com/facebook/metro/pull/821

Test Plan: - Added unit tests.

Reviewed By: robhogan

Differential Revision: D37161295

Pulled By: motiz88

fbshipit-source-id: 19af7091736ee34983d00e55d926662f923237b5