Fix a $FlowFixMe (#844)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/844

Removing the `$FlowFixMe: platform can be null` resulted in:

Cannot coerce `platform` to string because  null or undefined [1] should not be coerced.

The getAbsoluteAssetRecord method is called by various other methods in Assets.js. When researching whether the type for the platform parameter could be changed to a non-nullable string, I found that in upstream functions the platform can be identied from file sub-extensions (like `index.ios.js` as mentioned in `parsePlatformFilePath` for instance).

This means the platform parameter should indeed be optional, and therefore displaying it as '(unspecified)' when throwing an error seems reasonable as we can assume it will be present when it can be known.

Reviewed By: robhogan

Differential Revision: D37753953

fbshipit-source-id: 0a2acd87ccb5a8e7c311d41e5e7583be5dce21d2