Added isTurboModule helper function
Summary:
The function checks if a module extends TurboModule.

We are going to use this helper in a later diff to create the boundary schema.
The type schema is a representation of the module types. The boundary schema is a type schema for a module that is an interface usable by JavaScript code to call into native code and get results back(IE TurboModule spec).

We must check if the module has TurboModule spec to generate the entire schema. The boundary schema is an empty object if the module doesn't extend TurboModule because there is no interface for JavaScript code to call the native code. Therefore we can't generate the schema for the native code.

Reviewed By: GijsWeterings

Differential Revision: D38073929

fbshipit-source-id: 7b82a06f21652e6858bcfa220bd5b92191eec921