metro-react-native-interop-tools added getBoundarySchemaFromAST function
Summary:
The getBoundarySchemaFromAST function creates the main structure of our boundary schema. It contains just an InterfaceDeclarationTypeAnnotation(if there is one that extends TurboModule) because every NativeModule spec file must declare exactly one NativeModule interface.
Most parts of the type schema are the type annotations, getTypeAnnotation will be called by getBoundarySchemaFromAST to get the inner types annotation which will complete the entire type schema.
The getBoundarySchemaFromAST will be used by the version diffing, which gets two boundary schemas, and it checks if there are any incompatibilities between those schemas(for example, a function is renamed, or a parameter nullable in one schema and in the other schema is not).

Reviewed By: GijsWeterings

Differential Revision: D38278176

fbshipit-source-id: 09492554050e96418460567b881d1d24d076a554