metro-react-native-interop-tools added other annotation types
Summary:
I added the rest of the types for getTypeAnnotation as helpers because it's easier to understand and to test them with unit tests.
These functions remove other properties from the ast nodes, which are not helpful in creating the type schema.
Most parts of the type schema are type annotations, getTypeAnnotation will be called by the getBoundarySchemaFromAST function, which creates the main structure of our schema.

Reviewed By: arushikesarwani94

Differential Revision: D38385681

fbshipit-source-id: ed7e315e09496159e6bdd1b9a74f7f5966b3a446