metro-react-native-interop-tools added getObjectTypeAnnotation
Summary:
The getObjectTypeAnnotation function purpose is to get the type annotation for boundary schema from an BabelNodeObjectTypeAnnotation. This is a case of type annotation handled by the getTypeAnnotation function.
Most parts of the type schema are the type annotations, getTypeAnnotation will be called by getBoundarySchemaFromAST, which creates the main structure of our schema. Adding the type annotations will complete the entire type schema.
This function has two helpers, one for ObjectTypeProperty and one for ObjectTypeSpreadProperty. The helpers make the main function code more readable.
For now, it doesn't support object spread properties. We will add it later.

Reviewed By: arushikesarwani94

Differential Revision: D38252402

fbshipit-source-id: a868b3319c87739af104668702341e3732a24711