Separate in-memory graphs by resolver options
Summary:
Part of https://github.com/microsoft/rnx-kit/discussions/1803.

Ensures that the bundler will instantiate separate graphs for each combination of `customResolverOptions`. Once we pass them into the resolver (in a later diff), these options will be allowed to affect the output of the resolver and thus the contents of the graph, so we have to keep the graphs separate.

In practice there should not be a lot of duplication of the *same* bundle, even if an app makes lots of bundle requests sequentially with different resolver options - which I don't expect to be the case. Even in that case those bundles would generally not all be running and receiving HMR updates *at the same time*, so at least in principle we should be able to evict them from memory.[1]

[1] There may be followup work here to ensure this actually happens.

Reviewed By: jacdebug

Differential Revision: D38705297

fbshipit-source-id: 6436872521ccd864c122aa24a965c7b2e88c9cc9