metro-react-native-interop-tools modified basic types
Summary:
I split the BasicTypeAnnotation into more types and modified the structure of the getTypeAnnotation to be able to specifically take each type, in this way is more flexible and safe. I added BooleanLiteralTypeAnnotation, we can also have boolean literals.
Because we have all the types, I just changed the unknown type case by throwing an error which shows that our tool has a problem.

Reviewed By: GijsWeterings

Differential Revision: D38977909

fbshipit-source-id: 573c64ecbc8e5b3a12b6181ca5a9c3f79cdfc1ac