metro-react-native-interop-tools simplified the flag for function return type in the error helpers
Summary:
I changed the newVersion string variable to boolean isInFunctionReturn, now it is easier to understand that this variable keeps the state of the type, if it is in the function return type or not.
We are using this variable to create the error messages, because the left and right are switched in the function return comparison, we will still know which type is the new version and which is the old one.

Reviewed By: GijsWeterings

Differential Revision: D39465156

fbshipit-source-id: 39fe23afb281078d48e71d3b23218fe78923d367