Use read-only version of Graph<T> where possible
Summary:
Creates a new interface type, `ReadOnlyGraph<T>`, for use in functions that should only read from a Graph instance and not mutate it. In particular this includes the `customSerializer` and `experimentalSerializerHook` functions, which may be user-supplied via config.

This is not a breaking change since the serializer has always logically received a read-only graph. This diff merely codifies that in the Flow types.

Changelog: [Internal]

Reviewed By: huntie

Differential Revision: D40233735

fbshipit-source-id: d76482a9eaf344d6278430429eab6e9fef9a3627