Extract some crawl and watch management out of `index.js`
Summary:
Refactoring to begin the work of bringing crawl+watch closer together.

This simply pulls some code out of the bloated `index.js` into a `Watcher` class, which is responsible for managing crawl and watch backends.

The API is likely to change, but for now sticks to `crawl()` / `watch()` as separate instance methods (reflecting the current architecture) to minimise the churn in this diff.

Changelog: Internal

Reviewed By: motiz88

Differential Revision: D39891465

fbshipit-source-id: 8e3616847380bea7acf7117479a0a9a90526e19e