Refactor `metro-file-map`'s internal copy-on-change
Summary:
Currently, `metro-file-map` passes `InternalData` objects around from cache read, crawl, watch and then through each event-handling step (a promise chain). No file/map data is held as class instance state.

When we emit a `'change'` event to consumers, we emit (leak) references to the latest `data` values via `ModuleMap` and `HasteFS` objects, but at the same time we mark `mustCopy = true`.

Then, immediately prior to mutating data, if `mustCopy` we perform a (depth=1) copy and reassign `hasteMap`, so that the data already received by consumers doesn't change. Effectively, this means that consumers have a snapshot of the data, current as of the last event in each emitted batch.

## This change
Almost equivalently, rather than remembering when we have exposed data and copying it before mutating, we can flip this around - repeatedly mutate the same object, and only expose copies.

**Does this mean making copies more frequently / more than we have to?**
No -
 - In the new implementation, we explicitly create a new copy exactly once for each `'change'` (each batch of events accumulated over `CHANGE_INTERVAL` >= 30ms).
 - In the existing implementation, we set `mustCopy = true` (only) before every `'change'`, and we act on it before any mutation. Since every `'change'` emit implies a mutation has occurred since the last `'change'`, this also amounts to exactly one copy per `'change'` event emitted.

**Does this change the timing of emits?**
Copying this data structure is relatively expensive - ~200ms on a dev server on Meta's codebase - and synchronous. Doing this prior to `emit` simply delays the emit by that time. However, making a copy prior to mutating effectively delays events being added to the "ready to emit" queue by the same amount of time, so it amounts to the same experience.

**Is it even necessary to emit a snapshot?**
IMO, no - none of the consumers within Metro need a frozen copy of the data, and we can eliminate these from the event payload (and possibly move snapshot generation to an explicit API), but I'll work on that in diffs to follow. (In many cases, this 200ms is probably the slowest piece of the HMR process, and eliminating it might move us perceptibly closer to real-time.)

## Motivation
Symlink support requires a representation of the filesystem more complex than a flat map (ie, a graph/tree) - this would be more expensive to copy than a map, and somewhat expensive to instantiate from a serialised form. By taking more control over when we have to copy this data we'll be in a position to instantiate a tree representation no more often than necessary.

Changelog: Internal

Reviewed By: jacdebug

Differential Revision: D40829941

fbshipit-source-id: bf2947bec2dcb84fabf08004dba0b31df384403e