Make all getTransformOptions result properties optional
Summary:
Types `getTransformOptions` as returning an object with all-optional properties.

There's no need to force Metro config authors to return a bunch of empty properties when they just want to set `transform.inlineRequires` to `true`. (Also, as a practical matter, these are already treated as optional almost everywhere, including in the React Native template.)

In an upcoming diff I will document `getTransformOptions` on the Metro website and explain what all these options are for.

Changelog:

* **[Fix]:** Make all `getTransformOptions` result properties optional

Reviewed By: huntie

Differential Revision: D41400791

fbshipit-source-id: c30dfa57ec257f079fdb4e78863fdf0dc8f968d4