Document getTransformOptions
Summary:
We had a stub entry for `getTransformOptions` in the docs, but it contained zero information. This diff is an attempt to document it as-is; we'll likely want to remove/simplify some of this functionality later.

Closes https://github.com/facebook/metro/issues/845.

This diff also introduces some new(-to-Metro) formatting and styling to support the relatively complex content:

* A "deprecated" label, with CSS mostly lifted from the React Native website's implementation of platform labels.
  * The styling for `li .label` is all new.
* The use of a bulleted list to document a function's parameters.
* The use of nested bulleted lists to document complex object types.
* The use of Flow code blocks to show ancillary type definitions under the relevant config option.

Reviewed By: huntie

Differential Revision: D41400599

fbshipit-source-id: 061cc896fdb3fec9948c4045123d6c05e0d5cc81