Node crawler integration tests: Explicitly cover find vs recursive
Summary:
D42122025 adds an integration test for the "node crawler" that tests both values of `forceNodeFilesystemAPI`.

When `forceNodeFilesystemAPI` is `false`, the crawler runs a dummy `find` command to determine whether the system supports the necessary `find` syntax. If so, it proceeds with `find`, and if not it falls back to `fs` (the "Node filesystem API").

That left a potential gap in our tests, because if `hasNativeFindSupport` (eg because *it* has a bug or uses syntax supported nowhere) the `find` path would be broken, but the tests would still be green.

This diff separates `hasNativeFindSupport` into a separate file so it can be mocked, to force testing of the `find` path more explicitly. We'll need to skip the test on systems that don't support it, as we do with Watchman and (macOS) FSEvents.

NOTE: I'm sceptical about the `find` crawl, and may look to remove it. Although using "native" system features seems like a good idea, surprisingly some quick tests show it to be considerably slower than a recursive `fs` scan at all scales, at least on the OS/file systems I've tested.

Differential Revision: D42128545

fbshipit-source-id: bf5e61c6c463eaef5fa1fa9071c9ffaa831eb0eb