Introduce BundlerResolution type
Summary:
Changelog: Internal

## This diff

Creates a new object type, `BundlerResolution`, to represent the return value of the resolver as seen by the bundler (instead of returning just a single string value).

This is a mostly mechanical refactor, but there are some further changes in `ModuleResolution.js` where we remove the reliance on `getModule()` and the `ModuleT` generic (and for `sourceFile` resolutions, we now *directly* return the resolver's output).

The intention is to keep this close to the `Resolution` type that is in the resolver's public API (and in `resolveRequest` etc), and possibly remove the distinction between the two. *For now*, it's useful to continue to translate `assetFiles` and `empty` resolutions to `sourceFile` (inside `ModuleResolution.js`), without forcing all call sites to deal with the different resolution types.

## Next steps

In future work, `BundlerResolution` will carry more information that will allow the bundler to invalidate resolutions correctly (e.g. candidate paths that were attempted before the successful resolution).

Longer term, we might also want to use this structure to keep track of additional file dependencies that invalidate the *contents* of virtual modules (such as assets and `require.context` modules). Currently, asset modules are not invalidated correctly when a secondary asset file (e.g. a different scale variant of an image) is created/modified, and `require.context` invalidation is handled as a special case in DeltaBundler.

Reviewed By: robhogan

Differential Revision: D42442272

fbshipit-source-id: 12a57bb669791043755c39ac10a284e11678044e