Break up `InternalData` into `FileData`+`RawModuleMap` (1/2)
Summary:
In order to efficiently look up symlinks, we need a more complex structure than a simple `Map` to represent known files. The plan is to:
 1. Separate manipulation of the `files` Map from (Haste) `ModuleMap` data.
 2. Move access to (and more importantly, mutation of) the `files` map into the `HasteFS` class, implementing a `MutableFileSystem extends FileSystem` interface.
 3. Add an alternative `TreeFS` implementation of this interface, with an internal prefix-tree data structure, that we'll use when symlinks are enabled.
 4. If and when we have proven reliability and performance parity, remove `HasteFS`.

(Note that this also opens up the possibility of a "`LazyFS`" implementation, although for the moment the API will be synchronous)

This diff is the first part of that change - split the use of `InternalData`, which is essentially the raw data as read and written from the cache, into two objects - `FileData` and `RawModuleMap` (pre-existing types). Introduce `CacheData` to represent the combination of the two, which is now used only by cache APIs.

In the following diff, we'll replace the passing of `FileData` with an instance of `MutableFileSystem`.

Point of interest: I think this files/Haste separation is also what Christoph had in mind https://github.com/facebook/jest/pull/1427.

Changelog: [Internal]

Reviewed By: jacdebug

Differential Revision: D42303308

fbshipit-source-id: a3fa34311818ca3a0214687e6895e0bd2ba6c7b3