Mutate FileData through `MutableFileSystem` interface (2/2)
Summary:
In order to efficiently look up symlinks, we need a more complex structure than a simple `Map` to represent known files. The plan is to:
1. Separate manipulation of the `files` Map from (Haste) `ModuleMap` data. (D42303308 (https://github.com/facebook/metro/commit/ef9af830a61e4aa214ed321a1b15f6d5deb4b3eb))
 2. Move access to (and more importantly, mutation of) the `files` map into the `HasteFS` class, implementing a `MutableFileSystem extends FileSystem` interface.
 3. Add an alternative `TreeFS` implementation of this interface, with an internal prefix-tree data structure, that we'll use when symlinks are enabled.
 4. If and when we have proven reliability and performance parity, remove `HasteFS`.

This diff is part 2 - we introduce a new `MutableFileSystem` interface with APIs equivalent to the previous direct mutation of `FileData`, and encapsulate all mutation of the map (and everything in it) within `HasteFS`.

## Startup performance
Since we're introducing more indirection on some hot startup paths, we need to ensure there's no noticeable perf hit.

The figures in the test plan are single samples but they are consistent over multiple runs - this diff took a few draft iterations to get the speed almost up to the level of raw map manipulation. The key turned out to be minimising path normalisation, since `HasteFS` previously expected absolute paths for all of its APIs and had to normalise everything - it now accepts absolute or relative paths, and has shortcuts for fully normalised relative paths. The `bulkAddOrRemove` API takes `FileData` objects (implicitly pre-normalised, relative paths) and is therefore able to process much faster than multiple calls to `addOrRemove`.

Reviewed By: huntie

Differential Revision: D42344528

fbshipit-source-id: a47651154e18acb3490bb5158c40f88c2e365b8a