Gather symlink targets on crawl/change processing
Summary:
The groundwork has been done to store symlink targets in `metro-file-map`'s in-memory (and persisted) map, but we're not yet reading link targets.

This diff adds a `readLink` worker option and sets it when appropriate within `_processFile`, and delegates to the worker for consistency with other operations that require a "visit" to gather metadata subsequent to a crawl result or change event.

Symlink targets will then be stored for all symlinks in the cache, and we'll avoid any IO if their modification time doesn't change.

Changelog: [Internal]

Reviewed By: motiz88

Differential Revision: D42546808

fbshipit-source-id: 5c9737971f50dc90ade43f709ef896124023203e