Add `resolver.unstable_enableSymlinks` and remove symlinks+watchman restriction
Summary:
**Note: this change does not mean that symlinks are respected by the resolver (yet)!**

Add `resolver.unstable_enableSymlinks` (default: `false`) option to Metro config, and pass it though to the existing `enableSymlinks` option in `metro-file-map`.

At this point this is purely for ease of testing end-to-end behaviour as we incrementally change Fast Refresh and resolution to cope with symlinks.

Removes the Jest-inherited prohibition on using `enableSymlinks` with `useWatchman`, which is no longer necessary or relevant to the new implementation.

**Bikeshed: Why under `resolver`?**
Symlink support is mostly relevant to the resolver, in that if the resolver resolves to a file through a link, that fact is transparent to the rest of Metro. Third-party solutions for symlink-based projects have focused on the resolver. The fact that 90% of the work for this solution is within `metro-file-map` is an implementation detail of in-memory resolution. That said, we'll have an opportunity to revisit this when the config is promoted from `unstable_`

Changlog: [Experimental] - Add `resolver.unstable_enableSymlinks` config for first phase of symlink support.

Reviewed By: jacdebug

Differential Revision: D42552674

fbshipit-source-id: a6cfe38973b2a484af85f4a11274c4073710a883