Add integration test for require() + Fast Refresh
Summary:
Adds a test for the integration between Metro's `require()` and React, complementing the existing unit test. This lets us cover scenarios where the precise order of calling React's private APIs can impact the correctness of the result - e.g. if we call `performReactRefresh` before registering the latest versions of all components.

NOTE: The tests aren't comprehensive and the test harness may undergo changes. This is meant mostly as a blueprint for future tests accompanying further work on the `require()` implementation.

Reviewed By: GijsWeterings

Differential Revision: D42551229

fbshipit-source-id: e62f7c49061aab46ea56ee667a33c1f620f6e90b