DeltaCalculator - mark all graphs as dirty on any symlink change
Summary:
## Background
Related to the known issue with dependency graph correctness (where a new file that *should* take resolution precedence over an existing resolved dependency is ignored by the delta calculator), the issue with new or deleted symlinks is particularly acute. Symlinks to directories can make whole directory trees (dis)appear from a resolution perspective. In the wild, package managers like pnpm use symlinks to arrange `node_modules`, which is likely to create exactly the kind of resolution shadowing that causes correctness issues.

Moreover, symlinks are never in the dependency graph directly, because (even in a symlink-enabled world) all dependencies are resolved to real paths. This means there's nothing to be gained from trying to calculate a good incremental delta for a symlink change under the current model - it will always be empty, and sometimes incorrectly.

## Change
We have an ongoing project to resolve the dependency shadowing problem in general. In the mean time, this is a sledgehammer workaround to ensure correctness for symlink changes at the cost (only on symlink changes) of refresh performance.

This diff makes two changes:
 - On receipt of a symlink change event, the `DeltaCalculator` will set a marker such that the next `getDelta` reports all files in the graph as modified. This ensures that they are re-traversed and all resolutions reevaluated.
 - Ensure that the DeltaCalculator does not receive symlink change events from the watcher unless `unstable_enableSymlinks: true` - if symlinks are disabled, we won't resolve through them, so we shouldn't re-traverse the graph.

Reviewed By: jacdebug

Differential Revision: D42454225

fbshipit-source-id: f6d43f19c2b691ff49af9b5396e3b34c87a7a961