Process files before adding them to `FileSystem`
Summary:
`metro-file-map` takes raw crawl or watch results, "processes" them with a worker pool (compute hashes, read symlink targets), and makes metadata available to consumers.

Previously, files would be added to the file `Map` before processing, and then additional metadata would be set, and finally (when watching) an event would be emitted.

## Problem

This created a potential bug because files briefly exist, accessible in the `FileSystem` interface, in an intermediate state. Consider:
1. `foo.js` exists and has a SHA1 pre-calculated.
2. A change is made to `foo.js`, watcher backend emits a change event.
3. `FileSystem` is updated with the metadata of the unprocessed file, and an async worker starts processing.
4. Metro receives a bundle request for `foo.js`. It [fails hard](https://github.com/facebook/metro/blob/main/packages/metro/src/node-haste/DependencyGraph.js#L250) because the file has no calculated hash.
5. The worker completes and we populate the SHA1, and emit a 'change' event.

This also complicates `TreeFS`, because symlink targets
aren't guaranteed to be available to it.

## This diff

This diff switches things around so that the `FileSystem` implementation is only updated with ready-processed `FileMetaData`, and events are emitted to `metro-file-map` consumers synchronously with updates to `FileSystem`.

That means `FileSystem` will expose stale, complete state instead of fresher, incomplete state. Besides the fix above, this is okay and indeed preferable, because:
  - Two calls to `getSha1()` for the same file are guaranteed to return the same result unless there's been a 'change' event emitted on that file in the meantime. This is much more predictable.
   - In any case, any file system representation is *always* potentially stale - I/O isn't instant, OS events take time to propagate, so well behaved consumers should already be treating it as such.

Changelog: **[Fix]** Race condition where a very recently modified file might have missing metadata.

Reviewed By: jacdebug

Differential Revision: D42846676

fbshipit-source-id: b65a3907198fe616e957d21dbdd47a4365daa639