FileSystem: replace `getModifiedTime` and `getType` with `linkStats`
Summary:
`FileSystem` APIs such as `getSha1`, `getDependencies` are used outside `metro-file-map` and are normally expected to follow symlinks.

However, `getModifiedTime` is (only) used to determine whether to revisit a file after a change event on a real path, and is compared to `lstat().mtime`, so we don't want it to follow symlinks. Similarly, `getType` (which returns `?('f' | 'l')`) only make sense if symlinks are **not** followed.

Especially in implementing a symlink-enabled `FileSystem` it's clear that these APIs are inconsistent and potentially surprising.

This diff clarifies the API so that `linkStats` (like posix `lstat`)  returns information about a symlink, and all other APIs (like typical filesystem functions) should follow symlinks where the `FileSystem` API supports it.

Changelog: [Internal]

Reviewed By: huntie

Differential Revision: D43031551

fbshipit-source-id: bb0eeca0b6ad42c38882ef0d29f4aeb8849d929f