Watcher backends: Remove unused `hasIgnore`
Summary:
The various "ignore" mechanisms in `metro-file-map` are in a bit of tangle - we have explicit and implicit glob ignores, a set of regex arrays, and a `dot` parameter to exclude dotfiles.

This is a small step in the right direction by removing the `hasIgnore` variable that's always true for Metro, and has odd side effects when it's false.

Changelog: [Internal]

(There's no observable effect on Metro since we always set `ignorePattern`, and so `hasIgnore` is always `true`)

Reviewed By: motiz88

Differential Revision: D43234562

fbshipit-source-id: 166dca72774f3d6c3fb8c6825bf1a8d429439999