TreeFS: Symlink-enabled, prefix-tree implementation of `FileSystem` (#926)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/926

Currently, `HasteFS` is the only implementation of the `FileSystem` interface, used by Metro to check for file existence during resolution, and to retrieve metadata (file hashes, etc) during transformation. `HasteFS` is based on a simple, performant `Map` of  real paths to metadata.

This diff introduces `TreeFS`, which is based on an in-memory prefix tree representation, split on path separators, of all watched files.

By using a prefix tree, we can lookup files whose paths segments are directory symlinks or files which are themselves links to other files.

We use `TreeFS` in preference to `HasteFS` when `config.resolver.unstable_enableSymlinks` (default: `false`) is `true`.

Reviewed By: motiz88

Differential Revision: D42552999

fbshipit-source-id: 2705f34436d549831b51e4773bce39094a6ec507