Report errors in the console
Summary:
Error responses in CDP aren't generally displayed to the user. Currently, when metro-inspector-proxy fails to fetch a resource, we report success and embed the error message in the resource, allowing the user to see the error (e.g. if a source file failed to load, the debug client would still receive a success message, and display a "fake" source file containing the error).

This change makes metro-inspector-proxy report errors at the CDP layer (if applicable), and also send an error message to the console so the user can see it.

Reviewed By: newobj

Differential Revision: D43167737

fbshipit-source-id: e9ecf7f663e8f1aa8a6d48483bef0a0a7ca28ed8