Fix handling of deletion of a symlink pointing to a Haste module
Summary:
A bug I noticed while testing `unstable_enableSymlinks` on some Haste apps - because `FileSystem.getModuleName()` follows symlinks, we'd incorrectly clear a Haste module from the module map whenever a symlink pointing to the providing file was deleted.

This removes the use of `FileSystem.getModuleName()` and changes `MutableFileSystem.remove()` so that it returns the removed `FileMetadata`, if any. From that, we take the Haste ID (which is empty for symlinks) and remove the Haste module if that is populated.

Changelog:
[Experimental] `unstable_enableSymlinks`: Fix handling of deletion of a symlink pointing to a Haste module

Reviewed By: motiz88

Differential Revision: D43499958

fbshipit-source-id: ddff3d2236ccf7f60442156071289230ce9cec54