Add `unstable_preferTreeFS` to use TreeFS without enabling symlinks
Summary:
This allows the use of the new symlink-supporting prefix-tree `FileSystem` implementation, `TreeFS` *without* enabling symlinks.

The main purpose of this is to allow like-for-like performance comparisons of operations at this level - in particular, startup involving assembling the tree, and resolution under the new `exists` implementation.

Because `HasteFS` and `TreeFS` should have identical outputs on regular files, this setting is *not* a cache key component (unlike `resolver.unstable_enableSymlinks`).

Reviewed By: motiz88

Differential Revision: D43539702

fbshipit-source-id: 92639845f42d6c7a3d626de9853c9c360b45367a