Remove Lerna (#936)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/936

Lerna is powerful tool but we use very little of it. Now that `npm` natively supports `npm publish --workspaces` and `npm workspaces exec`, we actually don't need Lerna at all.

The version we have is old and we can't easily update with our "publish everything" release strategy. Removing it gets rid of a lot of legacy dependencies.

This removes 198 dev dependencies from `node_modules`, 20% of the lockfile.

Differential Revision: D43601474

fbshipit-source-id: 3d721684ce62596dbf813b70893a2cdb21df5d5f