Remove unused 'temp' dependency
Summary:
We only refer to `temp` in the context of mocking it in tests, which suggests it was at least a transitive dependency we needed to mock at some point - however it no longer is, as shown by the fact Yarn removes it from the lockfile.

This removes three packages (`temp`, `os-tmpdir`, `rimraf@2`) from Metro's prod dependencies.

Reviewed By: motiz88

Differential Revision: D43601352

fbshipit-source-id: 61d24675c8fece8f6f2defb64c18f261eb08e886