Fix minor typo metro-resolver warning
Summary:
Was seeing a frequent warning when testing fb-metro bundling a new app.

Attempted to import the module "{}" which is listed in the "exports" of "{}, however no match was resolved for thisrequest (platform = macos). Falling back to file-based resolution. Consider updating the call site or asking the package maintainer(s) to expose this API.

This fixes the typos in that warning.

Changelog: [Internal]

Reviewed By: robhogan

Differential Revision: D43849297

fbshipit-source-id: e65e68986ccb3506ebb2fb8eadfdabb56af0e1af