Remove `watcher.unstable_preferTreeFS` option, always use `TreeFS`, delete `HasteFS`
Summary:
## Context
`TreeFS` is necessary for Metro symlink support, but comes with a potential performance regression on startup and/or resolution.

We added the `unstable_preferTreeFS` config option to isolate and test this component on Meta's internal codebase, to quantify any perf regression.

## Findings
Happily, there's only a very small startup regression (<3ms on startup per 1k files on a mid-powered dev server), and approximately a 5% regression in net resolver performance, on the data gathered *before* the optimisations in D43774356.

With D43774356, the startup delay should reduce to about 1ms per 1000 files, and the net resolution regression should reduce to <2%.

We expect to recover this and more once we can drop `HasteFS` and use the capabilities of `TreeFS` to add more resolution-tailored APIs.

OTOH, `TreeFS` is already considerably more efficient for subtree operations like `require.context`, where `HasteFS` is forced to perform a full scan.

## This change
The performance of `TreeFS` is proven close enough to move everything over to it. `HasteFS` is then unreferenced outside tests, so we can delete it.

Reviewed By: jacdebug

Differential Revision: D43981991

fbshipit-source-id: cb9c9f2b710ed1352182bf873531d7800a580914